Kamar Alam Khan
This project is a web-based application designed to streamline the process of extracting data from images and PDFs of various document types. It leverages the power of Azure Form Recognizer, a Microsoft Azure AI service, to recognize and extract data. The extracted information is then presented to users in a structured tabular format.